# webdoc/public/openapi/queen-broker.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Queen MQ broker API",
    "version": "1.0.0",
    "summary": "The HTTP API of the Queen MQ broker.",
    "description": "Generated from the broker's own source at documentation build time by\n`webdoc/scripts/gen-openapi.mjs`: paths and methods come from the axum router,\nquery parameters from each handler's extractor, request bodies from the structs\nhandlers deserialize into, response codes from the `StatusCode` variants each\nhandler can return, and authorization from `auth::route_access_level`.\n\nResponse bodies on the hot paths are assembled by string concatenation rather than\nserialized from a type, and several management routes forward a stored procedure's\nJSON verbatim. Those responses carry `x-queen-schema: \"opaque\"` and are typed as an\nopen object: the field-by-field shape is documented on the reference pages instead of\nguessed at here.\n\nEvery 204 response has no body at all, not an empty JSON document.",
    "license": {
      "name": "Apache-2.0",
      "identifier": "Apache-2.0"
    }
  },
  "servers": [
    {
      "url": "http://localhost:6632",
      "description": "A broker with default settings"
    }
  ],
...
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Only enforced when JWT_ENABLED is on. Access levels are a role set, not a ladder: a write-only credential passes POST /api/v1/push and is rejected everywhere else."
      }
    },
...
  },
  "paths": {
    "/api/v1/push": {
      "post": {
        "operationId": "push",
        "tags": [
          "Message plane"
        ],
        "summary": "POST /api/v1/push",
        "parameters": [
          {
            "name": "x-queen-tenant",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Tenant to scope this request to. Read only when QUEEN_TENANCY_HEADER is on; unauthenticated by design, so the broker must not be reachable directly by tenants."
          }
        ],
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "description": "Shape not derivable from a Rust type. See the route's page under /reference/http for the field-by-field contract."
                }
              }
            },
            "x-queen-schema": "opaque"
          },
          "400": {
            "description": "Malformed request.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "x-queen-access-level": "write-only",
        "x-queen-tenant-scoped": true,
        "x-queen-handler": "server/src/handlers/data.rs::handle_push",
        "x-queen-query-params": "none",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "items": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "queue": {
                          "description": "Rust type `CowStr<'a>`."
                        },
                        "partition": {
                          "description": "Rust type `CowStr<'a>`."
                        },
                        "payload": {
                          "description": "Any JSON value."
                        },
                        "transactionId": {
                          "description": "Rust type `CowStr<'a>`."
                        }
                      },
                      "required": [
                        "queue",
                        "payload"
                      ]
                    }
                  }
                },
                "required": [
                  "items"
                ]
              }
            }
          },
          "x-queen-schema": "derived",
          "x-queen-schema-source": "PushBody"
        }
      }
    },
...
    "/api/v1/transaction": {
      "post": {
        "operationId": "transaction",
        "tags": [
          "Message plane"
        ],
        "summary": "POST /api/v1/transaction",
        "parameters": [
          {
            "name": "x-queen-tenant",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Tenant to scope this request to. Read only when QUEEN_TENANCY_HEADER is on; unauthenticated by design, so the broker must not be reachable directly by tenants."
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "description": "Shape not derivable from a Rust type. See the route's page under /reference/http for the field-by-field contract."
                }
              }
            },
            "x-queen-schema": "opaque"
          },
          "400": {
            "description": "Malformed request.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "x-queen-access-level": "read-write",
        "x-queen-tenant-scoped": true,
        "x-queen-handler": "server/src/handlers/data.rs::handle_transaction",
        "x-queen-query-params": "none",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "op": {
                    "description": "See the reference page for this route."
                  },
                  "operations": {
                    "description": "See the reference page for this route."
                  },
                  "requiredLeases": {
                    "description": "See the reference page for this route."
                  },
                  "type": {
                    "description": "See the reference page for this route."
                  },
                  "items": {
                    "description": "See the reference page for this route."
                  },
                  "transactionId": {
                    "description": "See the reference page for this route."
                  },
                  "partitionId": {
                    "description": "See the reference page for this route."
                  },
                  "consumerGroup": {
                    "description": "See the reference page for this route."
                  },
                  "status": {
                    "description": "See the reference page for this route."
                  },
                  "error": {
                    "description": "See the reference page for this route."
                  },
                  "leaseId": {
                    "description": "See the reference page for this route."
                  },
                  "value": {
                    "description": "See the reference page for this route."
                  },
                  "ok": {
                    "description": "See the reference page for this route."
                  },
                  "acks": {
                    "description": "See the reference page for this route."
                  },
                  "dlq": {
                    "description": "See the reference page for this route."
                  },
                  "group": {
                    "description": "See the reference page for this route."
                  },
                  "off": {
                    "description": "See the reference page for this route."
                  }
                },
                "additionalProperties": true
              }
            }
          },
          "x-queen-schema": "opaque",
          "description": "The handler reads this body as a dynamic JSON document, so the exact shape is not recoverable from a type. The listed keys are the ones the handler reads by name."
        }
      }
    },
...
    "/api/v1/configure": {
      "post": {
        "operationId": "configure",
        "tags": [
          "Queues"
        ],
        "summary": "POST /api/v1/configure",
        "parameters": [
          {
            "name": "x-queen-tenant",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Tenant to scope this request to. Read only when QUEEN_TENANCY_HEADER is on; unauthenticated by design, so the broker must not be reachable directly by tenants."
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "description": "Shape not derivable from a Rust type. See the route's page under /reference/http for the field-by-field contract."
                }
              }
            },
            "x-queen-schema": "opaque"
          },
          "400": {
            "description": "Malformed request.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "x-queen-access-level": "read-write",
        "x-queen-tenant-scoped": true,
        "x-queen-handler": "server/src/handlers/queues.rs::handle_configure",
        "x-queen-query-params": "none",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "queue": {
                    "description": "See the reference page for this route."
                  },
                  "options": {
                    "description": "See the reference page for this route."
                  },
                  "error": {
                    "description": "See the reference page for this route."
                  }
                },
                "additionalProperties": true
              }
            }
          },
          "x-queen-schema": "opaque",
          "description": "The handler reads this body as a dynamic JSON document, so the exact shape is not recoverable from a type. The listed keys are the ones the handler reads by name."
        }
      }
    },
    "/api/v1/resources/queues": {
      "get": {
        "operationId": "list_queues",
        "tags": [
          "Queues"
        ],
        "summary": "GET /api/v1/resources/queues",
        "parameters": [
          {
            "name": "x-queen-tenant",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Tenant to scope this request to. Read only when QUEEN_TENANCY_HEADER is on; unauthenticated by design, so the broker must not be reachable directly by tenants."
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "description": "Shape not derivable from a Rust type. See the route's page under /reference/http for the field-by-field contract."
                }
              }
            },
            "x-queen-schema": "opaque"
          },
          "500": {
            "description": "Internal error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "x-queen-access-level": "read-only",
        "x-queen-tenant-scoped": true,
        "x-queen-handler": "server/src/handlers/queues.rs::handle_list_queues",
        "x-queen-query-params": "none",
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/v1/stats/refresh": {
      "post": {
        "operationId": "stats_refresh",
        "tags": [
          "Status and metrics"
        ],
        "summary": "POST /api/v1/stats/refresh",
        "parameters": [],
        "responses": {
          "500": {
            "description": "Internal error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "x-queen-access-level": "admin",
        "x-queen-tenant-scoped": false,
        "x-queen-handler": "server/src/handlers/status.rs::handle_stats_refresh",
        "x-queen-query-params": "none",
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  }
}